Update C. Script - Data backup.sh

+++ 2. Linux Bash Task/C. Script - Data backup.sh
@@ -5,7 +5,7 @@ clear
 # Sometimes you need to pass a multiline text block to a command or variable. 
 # Bash has a special type of redirect called Here document (heredoc) that allows you to do that
 
-cat << EOF
+#cat << EOF
 # ---------------------------------------------------------------------------------
 #						C. Script - Data backup.sh
 # ---------------------------------------------------------------------------------
@@ -19,18 +19,18 @@ cat << EOF
 # [The command to run the script must be added to crontab 
 # with a run frequency of one minute] 
 # ---------------------------------------------------------------------------------
-EOF
-echo
+#EOF
+#echo
 
 
 # ==================================== FUNCTIONS ====================================
 
 # DATA info
 backupdate=$(date +%Y%m%d%H%M%S)
-echo "===================================================================="
-echo "Local TIME $backupdate"
-echo "===================================================================="
-echo
+#echo "===================================================================="
+#echo "Local TIME $backupdate"
+#echo "===================================================================="
+#echo
 
 
 Path_Config=".config_Script_Data_backup"
@@ -43,6 +43,8 @@ FILE=./$Path_Config/$Backup_FileName
 
 Line1_Backup_Settings="Path to source:"
 Line2_Backup_Settings="Path to destination:"
+
+
 
 func_Backup_and_LOG_File_Create() {
   echo "===================================================================="
@@ -139,18 +141,9 @@ func_check_sync_path_from2to() {
 
 func_copy_from2to() {
   #echo "----------func_copy_from2to------------------"
- 
+  echo "##################################################################" >> $for_log_file_info  
   echo "$LOGDATA == func_copy_from2to RUN" >> $for_log_file_info
   
-  path_FROM=$(cat $FILE | grep "$Line1_Backup_Settings" | sed 's/.*: //')
-  path_TO=$(cat $FILE | grep "$Line2_Backup_Settings" | sed 's/.*: //')
-  #echo $path_FROM
-  #echo $path_TO
-    
-  list_files_FROM=$(cat $FILE | grep "List files in TO path" | sed 's/.*: //')
-  list_files_TO=$(cat $FILE | grep "List files in FROM path" | sed 's/.*: //')
-  #echo $list_files_FROM
-  #echo $list_files_TO
   
   echo "---------------------------"
   echo "------ BEFORE Sync --------" 
@@ -159,29 +152,31 @@ func_copy_from2to() {
   
   echo "...... Folder FROM consist from such files:"
   # ls >> dir.txt
-  ls "$path_FROM" > $list_files_FROM
+  # ls "$path_FROM" > $list_files_FROM   # Перенес в блок сравнения CMP
   cat "$list_files_FROM"
   echo
   echo "...... In folder TO such files !!! WILL BE REMOVE !!!:"
-  ls "$path_TO" > $list_files_TO
+  #ls "$path_TO" > $list_files_TO        # Перенес в блок сравнения CMP
   cat "$list_files_TO"
   echo 
-  
-  #diff "$list_files_FROM" "$list_files_TO"  
-  # diff - compare files line by line
-  #    -y, --side-by-side / output in two columns  
+    
+  # diff "$list_files_FROM" "$list_files_TO"  
+  #     diff - compare files line by line
+  #            -y, --side-by-side / output in two columns  
   echo ":::::::::::::::::: DIFF between FROM & TO ::::::::::::::::::::::::"
   echo "::::::::::::::::::       BEFORE Sync      ::::::::::::::::::::::::"
   diff -y "$list_files_FROM" "$list_files_TO" > $Path_Config/changes.diff
   cat $Path_Config/changes.diff 
   echo
   
-  # Remove files in 
+  # Remove files in path TO and copy files from path FROM
+  #echo $path_FROM
+  #echo $path_TO
   rm -r "$path_TO/"
-  cp -r "$path_FROM/" "$path_TO/"
+  #cp -r "$path_FROM/" "$path_TO/"
+  cp -a "$path_FROM/." "$path_TO/"
   
-  
-  
+    
   echo "---------------------------"
   echo "------- AFTER Sync --------"
   echo "---------------------------"  
@@ -200,46 +195,65 @@ func_copy_from2to() {
   echo "::::::::::::::::::       AFTER Sync       ::::::::::::::::::::::::" 
   diff -y "$list_files_FROM" "$list_files_TO" > $Path_Config/changes.diff
   cat $Path_Config/changes.diff 
-  
+  sleep 1 #Добавил, т.к. было наложение следующей строчки в лог файле 
   echo "$LOGDATA == func_copy_from2to FINISH" >> $for_log_file_info
   echo
-  exit
-  
-  #if         
+  exit       
 }
 
 # ==================================== ========= ====================================
 
 # ====================================   MAIN    ====================================
 
-
-if [ -w "$FILE" ]; then
+if [ -w "$FILE" ]; then # Проверка наличия файла конфигурации.
   # При проверке существования файла наиболее часто используются операторы FILE -e и -f . 
   # Первый проверит, существует ли файл независимо от типа, а второй вернет истину, 
   # только если ФАЙЛ является обычным файлом (а не каталогом или устройством).
   # -e file    Проверяет, существует ли файл.
   # -f file    Проверяет, существует ли файл, и является ли он файлом.
   # -w file    Проверяет, существует ли файл, и доступен ли он для записи.
+  
   #echo "===================================================================="  
   #echo "$Backup_FileName exists, so Data Backup was configureted."
+      
+  func_check_sync_path_from2to #Проверка наличия директорий откуда и куда копировать
   
-  #echo "--------------------------------------------------------------------" 
+    
+  #echo "---------------------------------------------"
+  #echo "--- COMPARATE files in path FROM & TO  ------" 
+  #echo "---------------------------------------------" 
   
-  func_check_sync_path_from2to
+  path_FROM=$(cat $FILE | grep "$Line1_Backup_Settings" | sed 's/.*: //')
+  path_TO=$(cat $FILE | grep "$Line2_Backup_Settings" | sed 's/.*: //')
+  #echo $path_FROM
+  #echo $path_TO
+    
+  list_files_FROM=$(cat $FILE | grep "List files in TO path" | sed 's/.*: //')
+  list_files_TO=$(cat $FILE | grep "List files in FROM path" | sed 's/.*: //')
+  #echo $list_files_FROM
+  #echo $list_files_TO
   
-  
-  # Перенаправляем вывод в файл и показываем его на экране
-  #   Этот вариант подобен оператору > из предыдущего пункта, то имеется при записи в файл, 
-  #   все старые данные будут удалены. Если вам нужно дописать в файл, в конструкцию необходимо 
-  #   добавить параметр -a:
-  #          команда | tee -a /путь/к/файлу
-  
-  # Всю информацию с экрана о копировании файлов поместить в лог файл
-  func_copy_from2to | tee -a $for_log_file_info
-
-  
-  echo "!!!!!!!!!!!!!!! FINISh !!!!!!!!!!!!!!!!!!!!!!!!!"  
-  echo "===================================================================="
+    
+  ls "$path_FROM" > $list_files_FROM
+  ls "$path_TO" > $list_files_TO 
+  if cmp -s "$list_files_FROM" "$list_files_TO"
+  then
+     #echo "The files match"
+     exit
+  else
+     echo "The files are different"
+      # Перенаправляем вывод в файл и показываем его на экране
+      #   Этот вариант подобен оператору > из предыдущего пункта, то имеется при записи в файл, 
+      #   все старые данные будут удалены. Если вам нужно дописать в файл, в конструкцию необходимо 
+      #   добавить параметр -a:
+      #          команда | tee -a /путь/к/файлу
+      
+      # Всю информацию с экрана о копировании файлов поместить в лог файл
+      func_copy_from2to | tee -a $for_log_file_info
+  fi
+    
+  #echo "!!!!!!!!!!!!!!! FINISh !!!!!!!!!!!!!!!!!!!!!!!!!"  
+  #echo "===================================================================="
 else 
 
   func_Path_2Sync_Directories_Create
@@ -248,10 +262,32 @@ else
   echo "===================================================================="
   echo
 fi
-echo
+#echo
 
 # ==================================== ========= ====================================
 
+#------ V1:
+# sudo cp 'C. Script - Data backup.sh' /usr/local/bin/
+# sudo nano /etc/crontab
+# 0/1  *  * * * ubuntu    bash /usr/local/bin/script.sh
 
+#------ V2:
+# crontab -e
+# /home/ubuntu/L1_EPAM/2. Linux Bash Task/TO/C. Script - Gen_files_for_TO.sh
+# 1 * * * * sh './home/ubuntu/L1_EPAM/2. Linux Bash Task/C. Script - Gen_files_for_TO.sh'
+# 1 * * * * sh '/home/ubuntu/L1_EPAM/2.\ Linux\ Bash\ Task/C.\ Script\ -\ Data\ backup.sh'
+
+
+# crontab -l
+
+#    * * * * * Команда, которая будет выполнена
+#    - - - - -
+#    | | | | |
+#    | | | | - День недели (0 - 7) (воскресенье = 0 или 7)
+#    | | | --- Месяц (1 - 12)
+#    | | --- День месяца (1 - 31)
+#    | ---- Час (0 - 23)
+#    ----- Минута (0 - 59)
+#0/1 * * * * /usr/local/bin/script.sh
 
 # ==================================== ========= ====================================
